SLVSCODE-116 Defer call to listAllRules to when it is actually needed
In the current code, the promise evaluation is initiated as soon
as the tree data provider class is created. This is not needed
as long as the "SonarLint Rules" pane is closed. The proposed change
makes sure that the promise is lazily evaluated only the first time
its result is actually needed.